@@ -1,10 +1,10 @@
 # DEFINE FILES TO COMPILE TO TESTS.
-set(FILES "test.cpp")  
+set(FILES "test.cpp" "datatypes/fileStorage.cpp")  
 
 cmake_minimum_required(VERSION 3.14)
 project(AP_Email)
 
-set(CMAKE_CXX_STANDARD 14)
+set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
 include(FetchContent)
