@@ -3,9 +3,8 @@
 #include <vector>
 #include <optional>
 #include <memory>
-#include <Istorage.h>
-#include <fileStorage.h>
-#include <bloomFilterStorage.h>
+#include "datatypes/Istorage.h"
+#include "datatypes/fileStorage.h"
 
 // Should pass if test.cpp was compiled using the docker script.
 TEST(DockerTest, CompileFile) {
@@ -13,90 +12,109 @@ TEST(DockerTest, CompileFile) {
 }
 
 using namespace std;
+using namespace testing;
 
 // Concrete implementation for testing
 
 
 // Define test types
-bloomFilterStorage storage = new bloomFilterStorage();
+class fileStorageTest : public testing::Test {
+    protected:
+        unique_ptr<fileStorage> storage;
+    
+        void SetUp() override {
+            storage = make_unique<fileStorage>("test.txt");
+        }
+    
+        void TearDown() override {
+            storage->remove(); // clean up if necessary
+        }
+    };
+    
+// Test: constructor
+TEST_F(fileStorageTest, Constructor_CreatesFile) {
+    EXPECT_TRUE(filesystem::exists("../data/test.txt"));
+}
 
 // Test: save method
-TEST_F(storage, Save_ReturnsTrueOnSuccess) {
+    
+TEST_F(fileStorageTest, Save_ReturnsTrueOnSuccess) {
     string testData = "Hello, world!";
-    storage.save(testData);
-    result = storage.getUrls().load();
-    EXPECT_EQ(result, testData);
+    storage->save(testData);
+    auto result = storage->load();
+    ASSERT_TRUE(result.has_value());
+    EXPECT_EQ(result.value(), testData);
 }
 
 // Test: load method with no data
-TEST_F(storage, Load_ReturnsDataWhenDataExists) {
-    auto result = storage.load();
-    EXPECT_EQ(result, storage&);
+
+TEST_F(fileStorageTest, Load_returnsEmptyWhenNoData) {
+    storage->remove();
+    auto result = storage->load();
+    EXPECT_FALSE(result.has_value());
 }
 
-// Test: load method with data
-TEST_F(storage, Load_ReturnsDataWhenDataExists) {
+// Test: load method with specific data
+TEST_F(fileStorageTest, Load_ReturnsDataWhenDataExists) {
     string testData = "Hello, world!";
-    storage.save(testData);
-    
-    auto result = storage.load();
+    storage->save(testData);
+    auto result = storage->load(testData);
     ASSERT_TRUE(result.has_value());
     EXPECT_EQ(testData, result.value());
 }
 
 // Test: exists method with no data
-TEST_F(storage, Exists_ReturnsFalseWhenNoData) {
-    bool result = storage.exists();
-    EXPECT_FALSE(result);
+TEST_F(fileStorageTest, Exists_ReturnsFalseWhenNoData) {
+    bool result = storage->exists();
+    EXPECT_TRUE(result);
 }
 
-// Test: exists method with data
-TEST_F(storage, Exists_ReturnsTrueWhenDataExists) {
+// Test: exists method with specific data
+TEST_F(fileStorageTest, Exists_ReturnsTrueWhenDataExists) {
     string testData = "Hello, world!";
-    storage.save(testData);
-    
-    bool result = storage.exists();
+    storage->save(testData);
+    bool result = storage->exists(testData);
     EXPECT_TRUE(result);
 }
 
 // look at this and see if implementation really overwrites instead of appending
 
 // Test: overwrite data
-TEST_F(storage, Save_OverwritesExistingData) {
+TEST_F(fileStorageTest, Save_OverwritesExistingData) {
     string testData1 = "Hello, world!";
     string testData2 = "Goodbye, world!";
+    storage->save(testData1);
+    storage->save(testData2);
     
-    storage.save(testData1);
-    storage.save(testData2);
-    
-    auto result = storage.load();
+    auto result = storage->load();
     ASSERT_TRUE(result.has_value());
     EXPECT_EQ(testData2, result.value());
 }
 
-// Test with integers
-TEST_F(storage, CanStoreAndRetrieveIntegers) {
-    int testData = 42;
-    storage.save(testData);
+// // Test with integers
+// TEST(fileStorageTest, CanStoreAndRetrieveIntegers) {
+//     int testData = 42;
+//     storage->save(testData);
     
-    auto result = storage.load();
-    ASSERT_TRUE(result.has_value());
-    EXPECT_EQ(testData, result.value());
-}
+//     auto result = storage->load();
+//     ASSERT_TRUE(result.has_value());
+//     EXPECT_EQ(testData, result.value());
+// }
 
 // Test: remove method
-TEST_F(storage, Remove_DeletesData) {
+TEST_F(fileStorageTest, Remove_DeletesData) {
     string testData = "Hello, world!";
-    storage.save(testData);
+    storage->save(testData);
     
-    storage.remove(testData);
+    storage->remove(testData);
     
-    auto result = storage.loadUrls();
+    auto result = storage->load();
     EXPECT_FALSE(result.has_value());
 }
+
 // Test: remove method when no data exists
-TEST_F(storage, Remove_NoDataDoesNotThrow) {
-    EXPECT_NO_THROW(storage.remove());
+TEST_F(fileStorageTest, Remove_NoDataDoesNotThrow) {
+    EXPECT_NO_THROW(storage->remove());
 }
 
 
